+++ packages/core/src/index.ts
@@ -1,1 +1,2 @@
 export * from "./AgentForge.js";
+export * from "./AgentForgeState.js";